deprecate: leave re-export stub in utils/api.js pointing to apiBase.js
